DEV: Split tests for user preferences security page into individual file
Mixing multiple acceptance tests in a single file makes the acceptance
tests much harder to find.
