+++ app/assets/javascripts/discourse/tests/acceptance/preferences-test.js
@@ -1,6 +1,5 @@
 import {
   acceptance,
-  count,
   exists,
   query,
   updateCurrentUser,
@@ -569,56 +568,6 @@ acceptance("Ignored users", function (needs) {
     await visit(`/u/eviltrout/preferences/users`);
     updateCurrentUser({ moderator: true });
     assert.ok(exists(".user-ignore"), "it shows the list of ignored users");
-  });
-});
-
-acceptance("User Preferences - Security", function (needs) {
-  needs.user();
-  needs.pretender(preferencesPretender);
-
-  test("recently connected devices", async function (assert) {
-    await visit("/u/eviltrout/preferences/security");
-
-    assert.strictEqual(
-      query(
-        ".auth-tokens > .auth-token:nth-of-type(1) .auth-token-device"
-      ).innerText.trim(),
-      "Linux Computer",
-      "it should display active token first"
-    );
-
-    assert.strictEqual(
-      query(".pref-auth-tokens > a:nth-of-type(1)").innerText.trim(),
-      I18n.t("user.auth_tokens.show_all", { count: 3 }),
-      "it should display two tokens"
-    );
-    assert.strictEqual(
-      count(".pref-auth-tokens .auth-token"),
-      2,
-      "it should display two tokens"
-    );
-
-    await click(".pref-auth-tokens > a:nth-of-type(1)");
-
-    assert.strictEqual(
-      count(".pref-auth-tokens .auth-token"),
-      3,
-      "it should display three tokens"
-    );
-
-    const authTokenDropdown = selectKit(".auth-token-dropdown");
-    await authTokenDropdown.expand();
-    await authTokenDropdown.selectRowByValue("notYou");
-
-    assert.strictEqual(count(".d-modal:visible"), 1, "modal should appear");
-
-    await click(".modal-footer .btn-primary");
-
-    assert.strictEqual(
-      count(".pref-password.highlighted"),
-      1,
-      "it should highlight password preferences"
-    );
   });
 });
